for Pavel

+++ app/Http/Controllers/Forum/CommentController.php
@@ -51,19 +51,19 @@ class CommentController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(ForumCommentCreateRequest $request)
     {
         $data = $request->input();
-
         $item = (new ForumComment())->create($data);
 
         if ($item) {
+
             return back()
-                ->withErrors(['msg' => 'Ошибка сохранения'])
                 ->with(['success' => 'Успешно добавленно']);
         } else {
+
             return back()
                 ->withErrors(['msg' => 'Ошибка сохранения'])
                 ->withInput();
@@ -97,49 +97,60 @@ class CommentController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(ForumCommentUpdateRequest $request, $id)
     {
-        $item = $this->forumCommentRepository->getEdit($id);
-        if ($item->user_id != Auth::user()->id) {
-            return  abort(403, 'Доступ запрещен');
-        }
-        if (empty($item)) {
-            return back()
-                ->withErrors(['msg' => "Запись id=[{$id}] не найдена"])
-                ->withInput();
+        if ($this->forumCommentRepository->Owner($id)) {
+            $item = $this->forumCommentRepository->getEdit($id);
+
+            if (empty($item)) {
+
+                return back()
+                    ->withErrors(['msg' => "Запись id=[{$id}] не найдена"])
+                    ->withInput();
+            }
+
+            $data = $request->all();
+            $result = $item->update($data);
+
+            if ($result) {
+
+                return back()
+                    ->with(['success' => 'Успешно сохранено']);
+            } else {
+
+                return back()
+                    ->withErrors(['msg' => 'Ошибка сохранения'])
+                    ->withInput();
+            }
         }
 
-        $data = $request->all();
-
-        $result = $item->update($data);
-
-        if ($result) {
-            return redirect()
-                ->back()
-                ->with(['success' => 'Успешно сохранено']);
-        } else {
-            return back()
-                ->withErrors(['msg' => 'Ошибка сохранения'])
-                ->withInput();
-        }
+        return abort(403, 'Доступ запрещен');
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy($id)
     {
-        $result = ForumComment::destroy($id);
+        if ($this->forumCommentRepository->Owner($id)) {
+            $result = ForumComment::destroy($id);
 
-        if ($result) {
-            return back()->with(['success' => 'Запись удалена']);
+            if ($result) {
+
+                return back()
+                    ->with(['success' => 'Запись удалена']);
+            } else {
+
+                return back()
+                    ->withErrors(['msg' => 'Ошибка удаления']);
+            }
         } else {
-            return back()->withErrors(['msg' => 'Ошибка удаления']);
+            return abort(403, 'Доступ запрещен');
         }
     }
 }
